# contracts/v2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SaturnPath V2 API",
    "version": "0.1.0",
    "description": "Authoritative contract for the SaturnPath native iOS app and V2 admin console. Student question payloads never contain answer keys before submission."
  },
  "servers": [
    { "url": "https://saturnpath.app/api/v2", "description": "Production" },
    { "url": "http://localhost:3000/api/v2", "description": "Local development" }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/home": {
      "get": {
        "operationId": "getHome",
        "tags": ["Bootstrap"],
        "responses": {
          "200": { "$ref": "#/components/responses/HomeResponse" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase Auth access token. Server secrets are never accepted from clients."
      }
    },
...
    "responses": {
...
      "HomeResponse": { "description": "Home state", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Home" } } } },
...
      "Unauthorized": { "$ref": "#/components/responses/Error401" },
...
      "Error401": { "description": "Authentication required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
...
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "requestId"],
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "requestId": { "type": "string" },
          "fieldErrors": { "type": "object", "additionalProperties": { "type": "array", "items": { "type": "string" } } }
        }
      },
...
      "Ring": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "label", "completedValue", "targetValue", "progress"],
        "properties": {
          "id": { "type": "string", "enum": ["score_growth", "weakness_removed", "time_saved"] },
          "label": { "type": "string" },
          "completedValue": { "type": "number", "minimum": 0 },
          "targetValue": { "type": "number", "exclusiveMinimum": 0 },
          "progress": { "type": "number", "minimum": 0, "maximum": 1 },
          "unit": { "type": "string" }
        }
      },
      "Home": {
        "type": "object",
        "additionalProperties": false,
        "required": ["rings", "targetScore", "minutesSaved", "recommendation"],
        "properties": {
          "scoreEstimate": { "type": ["integer", "null"], "minimum": 400, "maximum": 1600 },
          "scoreEstimateLow": { "type": ["integer", "null"] },
          "scoreEstimateHigh": { "type": ["integer", "null"] },
          "targetScore": { "type": "integer", "minimum": 400, "maximum": 1600 },
          "testDate": { "type": ["string", "null"], "format": "date" },
          "minutesSaved": { "type": "integer", "minimum": 0 },
          "rings": { "type": "array", "minItems": 3, "maxItems": 3, "items": { "$ref": "#/components/schemas/Ring" } },
          "recommendation": { "$ref": "#/components/schemas/DailyRecommendation" }
        }
      },
      "DailyRecommendation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "date", "totalMinutes", "newMinutes", "reviewMinutes", "whySelected"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "date": { "type": "string", "format": "date" },
          "totalMinutes": { "type": "integer", "minimum": 0 },
          "newMinutes": { "type": "integer", "minimum": 0 },
          "reviewMinutes": { "type": "integer", "minimum": 0 },
          "whySelected": { "type": "array", "minItems": 1, "maxItems": 3, "items": { "type": "string" } }
        }
      },
...
    }
  }
}